@@ -42,63 +43,62 @@
     "sharp": "^0.32.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@babel/plugin-transform-modules-commonjs": "^7.18.6",
+    "@babel/core": "^7.21.4",
+    "@babel/plugin-transform-modules-commonjs": "^7.21.2",
     "@babel/preset-react": "^7.18.6",
-    "@mdx-js/mdx": "^2.1.3",
-    "@types/body-scroll-lock": "^2.6.1",
+    "@mdx-js/mdx": "^2.3.0",
+    "@types/body-scroll-lock": "^3.1.0",
     "@types/classnames": "^2.2.10",
     "@types/debounce": "^1.2.1",
-    "@types/github-slugger": "^1.3.0",
-    "@types/mdx-js__react": "^1.5.2",
-    "@types/node": "^14.6.4",
+    "@types/mdx-js__react": "^1.5.5",
+    "@types/node": "^18.15.11",
     "@types/parse-numeric-range": "^0.0.1",
-    "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.5",
-    "@typescript-eslint/eslint-plugin": "^5.36.2",
-    "@typescript-eslint/parser": "^5.36.2",
+    "@types/react": "^18.0.37",
+    "@types/react-dom": "^18.0.11",
+    "@typescript-eslint/eslint-plugin": "^5.59.0",
+    "@typescript-eslint/parser": "^5.59.0",
     "asyncro": "^3.0.0",
-    "autoprefixer": "^10.4.2",
+    "autoprefixer": "^10.4.14",
     "babel-eslint": "10.x",
     "eslint": "7.x",
     "eslint-config-next": "^13.3.0",
-    "eslint-config-react-app": "^5.2.1",
+    "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-flowtype": "4.x",
     "eslint-plugin-import": "2.x",
     "eslint-plugin-jsx-a11y": "6.x",
-    "eslint-plugin-react": "7.x",
+    "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^0.0.0-experimental-fabef7a6b-20221215",
-    "fs-extra": "^9.0.1",
-    "globby": "^11.0.1",
-    "gray-matter": "^4.0.2",
-    "husky": "^7.0.4",
+    "fs-extra": "^11.1.1",
+    "globby": "^13.1.4",
+    "gray-matter": "^4.0.3",
+    "husky": "^8.0.3",
     "is-ci": "^3.0.1",
-    "lint-staged": ">=10",
-    "mdast-util-to-string": "^1.1.0",
-    "metro-cache": "0.72.2",
+    "lint-staged": "^13.2.1",
+    "mdast-util-to-string": "^3.2.0",
+    "metro-cache": "^0.76.2",
     "npm-run-all": "^4.1.5",
-    "patch-package": "^6.2.2",
-    "postcss": "^8.4.5",
-    "postcss-flexbugs-fixes": "4.2.1",
-    "postcss-preset-env": "^6.7.0",
-    "prettier": "^2.5.1",
-    "reading-time": "^1.2.0",
-    "remark": "^12.0.1",
-    "remark-external-links": "^7.0.0",
-    "remark-html": "^12.0.0",
-    "remark-images": "^2.0.0",
-    "remark-slug": "^7.0.0",
-    "remark-unwrap-images": "^2.0.0",
-    "retext": "^7.0.1",
-    "retext-smartypants": "^4.0.0",
+    "patch-package": "^6.5.1",
+    "postcss": "^8.4.22",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-preset-env": "^8.3.2",
+    "prettier": "^2.8.7",
+    "reading-time": "^1.5.0",
+    "remark": "^14.0.2",
+    "remark-external-links": "^9.0.1",
+    "remark-html": "^15.0.2",
+    "remark-images": "^3.1.0",
+    "remark-slug": "^7.0.1",
+    "remark-unwrap-images": "^3.0.1",
+    "retext": "^8.1.0",
+    "retext-smartypants": "^5.2.0",
     "rss": "^1.2.2",
     "tailwindcss": "^3.3.1",
-    "typescript": "^4.0.2",
-    "unist-util-visit": "^2.0.3",
-    "webpack-bundle-analyzer": "^4.5.0"
+    "typescript": "^5.0.4",
+    "unist-util-visit": "^4.1.2",
+    "webpack-bundle-analyzer": "^4.8.0"
   },
   "engines": {
-    "node": ">=12.x"
+    "node": ">=16.x"
   },
   "nextBundleAnalysis": {
     "budget": null,
