@@ -1,15 +1,8 @@
-/*!
- * Based on 'gatsby-remark-autolink-headers'
- * Original Author: Kyle Mathews <mathews.kyle@gmail.com>
- * Updated by Jared Palmer;
- * Copyright (c) 2015 Gatsbyjs
- */
+import {toString} from 'mdast-util-to-string';
+import {visit} from 'unist-util-visit';
+import GithubSlugger from 'github-slugger';
 
-const toString = require('mdast-util-to-string');
-const visit = require('unist-util-visit');
-const toSlug = require('github-slugger').slug;
-
-function patch(context, key, value) {
+function patch(context: any, key: any, value: any) {
   if (!context[key]) {
     context[key] = value;
   }
@@ -18,8 +11,8 @@ function patch(context, key, value) {
 
 const svgIcon = `<svg aria-hidden="true" height="16" version="1.1" viewBox="0 0 16 16" width="16"><path fill-rule="evenodd" d="M4 9h1v1H4c-1.5 0-3-1.69-3-3.5S2.55 3 4 3h4c1.45 0 3 1.69 3 3.5 0 1.41-.91 2.72-2 3.25V8.59c.58-.45 1-1.27 1-2.09C10 5.22 8.98 4 8 4H4c-.98 0-2 1.22-2 2.5S3 9 4 9zm9-3h-1v1h1c1 0 2 1.22 2 2.5S13.98 12 13 12H9c-.98 0-2-1.22-2-2.5 0-.83.42-1.64 1-2.09V6.25c-1.09.53-2 1.84-2 3.25C6 11.31 7.55 13 9 13h4c1.45 0 3-1.69 3-3.5S14.5 6 13 6z"></path></svg>`;
 
-module.exports = ({icon = svgIcon, className = `anchor`} = {}) => {
-  return function transformer(tree) {
+export default ({icon = svgIcon, className = `anchor`} = {}) => {
+  return function transformer(tree: any) {
     const ids = new Set();
     visit(tree, 'heading', (node) => {
       let children = [...node.children];
@@ -30,24 +23,22 @@ module.exports = ({icon = svgIcon, className = `anchor`} = {}) => {
         const isValidCustomId = id.startsWith('/*') && id.endsWith('*/');
         if (!isValidCustomId) {
           throw Error(
-            'Expected header ID to be like: {/*some-header*/}. ' +
-              'Instead, received: ' +
-              id
+            `Expected header ID to be like: {/*some-header*/}. Instead, received: ${id}`
           );
         }
         id = id.slice(2, id.length - 2);
-        if (id !== toSlug(id)) {
+        const slugId = new GithubSlugger().slug(id);
+
+        if (id !== slugId) {
           throw Error(
-            'Expected header ID to be a valid slug. You specified: {/*' +
-              id +
-              '*/}. Replace it with: {/*' +
-              toSlug(id) +
-              '*/}'
+            `Expected header ID to be a valid slug. You specified: {/*${id}*/}.
+            Replace it with: {/*${slugId}*/}`
           );
         }
       } else {
         // # My header
-        id = toSlug(toString(node));
+        const slugId = new GithubSlugger().slug(toString(node));
+        id = slugId;
       }
 
       if (ids.has(id)) {
