@@ -1,21 +1,21 @@
-const visit = require('unist-util-visit');
-const retext = require('retext');
-const smartypants = require('retext-smartypants');
+import {visit} from 'unist-util-visit';
+import {retext} from 'retext';
+import smartypants from 'retext-smartypants';
 
-function check(parent) {
+function check(parent: any) {
   if (parent.tagName === 'script') return false;
   if (parent.tagName === 'style') return false;
   return true;
 }
 
-module.exports = function (options) {
+export default function (options: any) {
   const processor = retext().use(smartypants, options);
 
-  function transformer(tree) {
-    visit(tree, 'text', (node, index, parent) => {
+  function transformer(tree: any) {
+    visit(tree, 'text', (node, _, parent) => {
       if (check(parent)) node.value = String(processor.processSync(node.value));
     });
   }
 
   return transformer;
-};
+}
