@@ -1,4 +1,4 @@
-import tailwindConfig from '../../../../tailwind.config';
+import tailwindConfig from '../../../../tailwind.config.cjs';
 
 export const CustomTheme = {
   colors: {
